+++ atompack/__init__.py
@@ -1,4 +1,0 @@
-from atompack.atom import Atom
-from atompack.crystal import Crystal, metric_tensor
-from atompack.molecule import Molecule
-from atompack.structure import Structure