+++ atompack/atom.py
@@ -1,41 +1,33 @@
-from typing import Optional
+import copy
 
 import numpy as np
 
+from atompack.util import AsDict
 
-class Atom(object):
-    """Representation of an atom with dynamic attributes.
+
+class Atom(AsDict):
+    """Container to store metadata about a single atom.
     
     Notes:
         Any `kwargs` passed to `__init__()` are dynamically set as instance variables.
-        This enables arbitrary atomic data to be associated with each instance in an
-        elegant and pythonic way. `position` and `symbol` are left as optional arguments
-        with default values so that each instance is guaranteed to have them as 
-        attributes for use in internal operations.
 
     Args:
-        position: Location vector in 3D cartesian space.
-            Mutating `position` is a logical error if the change results in atoms 
-            overlapping or existing out of bounds.
-        symbol: IUPAC chemical symbol.
+        position: Position vector in 3D cartesian space.
 
     Example:
-        >>> from atompack import Atom
+        >>> from atompack.atom import Atom
         >>> import numpy as np
         >>> 
-        >>> atom = Atom(charge=-2)
-        >>>
-        >>> assert np.array_equal(atom.position, np.zeros(3))
-        >>> assert atom.symbol == "Undefined"
+        >>> atom = Atom(np.zeros(3), charge=-2)
         >>> assert atom.charge == -2
     """
 
-    def __init__(self, position: Optional[np.ndarray] = None, symbol: Optional[str] = None, **kwargs) -> None:
-        if position is None:
-            position = np.zeros(3)
-        self.position = position
-        if symbol is None:
-            symbol = "Undefined"
-        self.symbol = symbol
+    def __init__(self, position: np.ndarray, **kwargs) -> None:
         for k, v in kwargs.items():
             setattr(self, k, v)
+        self._position = position
+
+    @property
+    def position(self) -> np.ndarray:
+        """Returns a copy of the atom's position vector."""
+        return copy.deepcopy(self._position)