+++ CHANGELOG.md
@@ -5,12 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.2] - ???
+## [0.3.0] - 2020-09-08
 
 ### Added
 
-* `Molecule` class.
+* `bond` module.
+* `molecule` module.
+* `topology` module.
+* `crystal.UnitCell` class with many subclasses.
 * Support for `pipenv` installation.
+* Additional benchmark tests.
+
+### Changed
+
+* `crystal.Crystal` now inherits from `topology.Topology`.
+
+### Removed
+
+* C extensions
+* `structure` module.
 
 ## [0.2.1] - 2020-08-02